+++ scripts/tts_dataset_files/wordid_to_nemo_cmu.tsv
@@ -1,0 +1,327 @@
+word_id	nemo_ipa_cmu
+abstract_adj-nou	ˈæbˌstɹækt
+abstract_vrb	æbˈstɹækt
+abuse_nou	əbˈjus
+abuse_vrb	əbˈjuz
+abuses_nou	əbˈjusɪz
+abuses_vrb	əbˈjuzɪz
+addict_nou	ˈæˌdɪkt
+addict_vrb	əˈdɪkt
+advocate_nou	ˈædvəkət
+advocate_vrb	ˈædvəˌkeɪt
+affect_nou-psy	ˈæˌfɛkt
+affect	əˈfɛkt
+affiliate_nou	əˈfɪɫiət
+affiliate_vrb	əˈfɪɫiˌeɪt
+aged_adj	ˈeɪdʒɪd
+aged	ˈeɪdʒd
+aggregate_adj-nou	ˈæɡɹəɡət
+aggregate_vrb	ˈæɡɹəɡeɪt
+alternate_adj-nou	ˈɔɫtɝnət
+alternate_vrb	ˈɔɫtɝˌneɪt
+analyses_nou	æˈnæɫɪˌsiz
+analyses_vrb	ænəˈɫaɪˌzɪz
+animate_adj-nou	ˈænəmət
+animate_vrb	ˈænəˌmeɪt
+appropriate_adj	əˈpɹoʊpɹiət
+appropriate_vrb	əˈpɹoʊpɹiˌeɪt
+approximate_adj-nou	əˈpɹɑksəmət
+approximate_vrb	əˈpɹɑksəˌmeɪt
+articulate_adj	ɑɹˈtɪkjəɫət
+articulate_vrb	ɑɹˈtɪkjəˌɫeɪt
+associate_adj-nou	əˈsoʊsiət
+associate_vrb	əˈsoʊsiˌeɪt
+attribute_nou	ˈætɹəbˌjut
+attribute_vrb	əˈtɹɪbˌjut
+august_adj	ˌɑˈɡəst
+august	ˈɑɡəst
+august_nam	ˈaʊˌɡʊst
+axes_nou	ˈækˌsiz
+axes_nou-vrb	ˈækˌsɪz
+bass	ˈbeɪs
+bass_corp	ˈbæs
+blessed_adj	ˈbɫɛsəd
+blessed_vrb	ˈbɫɛst
+bologna_geo	bəˈɫoʊˌŋɑ
+bologna	bəˈɫoʊni
+bow_nou-knot	ˈboʊ
+bow_nou-ship	ˈbaʊ
+buffet_nou	bəˈfeɪ
+buffet_vrb	ˈbəfət
+celtic_adj-nou-sports	ˈsɛɫtɪk
+celtic	ˈkɛɫtɪk
+close_adj-nou	ˈkɫoʊs
+close_vrb	ˈkɫoʊz
+combine_nou	ˈkɑmbaɪn
+combine_vrb	kəmˈbaɪn
+compact_adj-nou	ˈkɑmpækt
+compact_vrb	kəmˈpækt
+compound_nou	ˈkɑmpaʊnd
+compound_vrb	kəmˈpaʊnd
+compress_nou	ˈkɑmpɹɛs
+compress	kəmˈpɹɛs
+conduct_nou	ˈkɑndəkt
+conduct_vrb	kɑnˈdəkt
+confines_nou	ˈkɑnˌfaɪnz
+confines_vrb	kənˈfaɪnz
+conflict_nou	ˈkɑnfɫɪkt
+conflict_vrb	kənˈfɫɪkt
+conglomerate_adj-nou	kənˈɡɫɑmɝət
+conglomerate_vrb	kənˈɡɫɑmɝˌeɪt
+conjugate_adj-nou	ˈkɑndʒəˌɡeɪt
+conjugate_vrb	ˈkɑndʒəˌɡeɪt
+conscript_nou	ˈkɑnsˌkɹɪpt
+conscript_vrb	kənsˈkɹɪpt
+console_nou	ˈkɑnsoʊɫ
+console_vrb	kənˈsoʊɫ
+consort_nou	ˈkɑnˌsɔɹt
+consort_vrb	kənˈsɔɹt
+construct_nou	ˈkɑnstɹəkt
+construct_vrb	kənˈstɹəkt
+consummate_adj	ˈkɑnsəmət
+consummate_vrb	ˈkɑnsəˌmeɪt
+content_adj-nou-vrb	kənˈtɛnt
+content_nou	ˈkɑntɛnt
+contest_nou	ˈkɑntɛst
+contest_vrb	kənˈtɛst
+contract_nou	ˈkɑnˌtɹækt
+contract_vrb	kənˈtɹækt
+contrast_nou	ˈkɑntɹæst
+contrast_vrb	kənˈtɹæst
+converse_adj-nou	ˈkɑnvɝs
+converse_vrb	kənˈvɝs
+convert_nou	ˈkɑnvɝt
+convert_vrb	kənˈvɝt
+convict_nou	ˈkɑnvɪkt
+convict_vrb	kənˈvɪkt
+coordinate_adj-nou	koʊˈɔɹdənət
+coordinate_vrb	koʊˈɔɹdəˌneɪt
+correlate_nou	ˈkɔɹəɫət
+correlate_nou-vrb	ˈkɔɹəˌɫeɪt
+decrease_nou	ˈdiˌkɹis
+decrease_vrb	dɪˈkɹis
+defect_nou	ˈdifɛkt
+defect_vrb	dɪˈfɛkt
+degenerate_adj-nou	dɪˈdʒɛnɝət
+degenerate_vrb	dɪˈdʒɛnɝˌeɪt
+delegate_nou	ˈdɛɫəɡət
+delegate_vrb	ˈdɛɫəˌɡeɪt
+deliberate_adj	dɪˈɫɪbɝət
+deliberate_vrb	dɪˈɫɪbɝˌeɪt
+desert_nou	ˈdɛzɝt
+desert_vrb	dɪˈzɝt
+deviate_nou	ˈdiˌviət
+deviate_vrb	ˈdiviˌeɪt
+diagnoses_nou	ˌdaɪəɡˈnoʊsiz
+diagnoses_vrb	ˌdaɪəɡˈnoʊsəz
+diffuse_adj	dɪfˈjus
+diffuse_vrb	dɪfˈjuz
+discard_nou	ˈdɪsˌkɑɹd
+discard_vrb	dəsˈkɑɹd
+discharge_nou	ˈdɪsˌtʃɑɹdʒ
+discharge_vrb	dɪsˈtʃɑɹdʒ
+discount_nou	ˈdɪskaʊnt
+discount_vrb	dɪˈskaʊnt
+document_nou	ˈdɑkjumɛnt
+document_vrb	ˈdɑkjəmɛnt
+dove	ˈdəv
+dove_vrb	ˈdoʊv
+duplicate_adj-nou	ˈdupɫəkət
+duplicate_vrb	ˈdupɫəˌkeɪt
+elaborate_adj	ɪˈɫæbɹət
+elaborate_vrb	ɪˈɫæbɝˌeɪt
+entrance_nou	ˈɛntɹəns
+entrance_vrb	əˈntɹæns
+escort_nou	ˈɛskɔɹt
+escort_vrb	ɛˈskɔɹt
+estimate_nou	ˈɛstəmət
+estimate_vrb	ˈɛstəˌmeɪt
+excuse_nou	ɪkˈskjus
+excuse_vrb	ɪkˈskjuz
+expatriate_nou	ɛkˈspeɪtɹiət
+expatriate_vrb	ɛkˈspeɪtɹiˌeɪt
+exploit_nou	ˈɛksˌpɫɔɪt
+exploit_vrb	ˌɛksˈpɫɔɪt
+export_nou	ˈɛkspɔɹt
+export_vrb	əkˈspɔɹt
+expose_nou	ˌɛkˌspoʊˈzeɪ
+expose_vrb	əkˈspoʊz
+extract_nou	ˈɛkˌstɹækt
+extract_vrb	ɪkˈstɹækt
+fragment_nou	ˈfɹæɡmənt
+fragment_vrb	ˌfɹæɡˈmɛnt
+frequent_adj	ˈfɹikwənt
+frequent_vrb	ˌfɹiˈkwɛnt
+graduate_adj-nou	ˈɡɹædʒuwət
+graduate_vrb	ˈɡɹædʒuˌeɪt
+house_nou	ˈhaʊs
+house_vrb	ˈhaʊz
+impact_nou	ˈɪmpækt
+impact_vrb	ˌɪmˈpækt
+implant_nou	ˈɪmˌpɫænt
+implant_vrb	ˌɪmˈpɫænt
+implement_nou	ˈɪmpɫəmənt
+implement_vrb	ˈɪmpɫəˌmɛnt
+import_nou	ˈɪmˌpɔɹt
+import_vrb	ˌɪmˈpɔɹt
+incense_nou	ˈɪnˌsɛns
+incense_vrb	ˌɪnˈsɛns
+incline_nou	ˈɪnkɫaɪn
+incline_vrb	ˌɪnˈkɫaɪn
+increase_nou	ˈɪnˌkɹis
+increase_vrb	ˌɪnˈkɹis
+increment_nou	ˈɪnkɹəmənt
+increment_vrb	ˈɪnkɹəˌmɛnt
+initiate_nou	əˈnɪˌʃiət
+initiate_vrb	ˌɪˈnɪʃiˌeɪt
+insert_nou	ˈɪnˌsɝt
+insert_vrb	ˌɪnˈsɝt
+instrument_nou	ˈɪnstɹəmənt
+instrument_vrb	ˈɪnstɹəˌmɛnt
+insult_nou	ˈɪnˌsəɫt
+insult_vrb	ˌɪnˈsəɫt
+interchange_nou	ˈɪntɝˌtʃeɪndʒ
+interchange_vrb	ˌɪntɝˈtʃeɪndʒ
+intimate_adj	ˈɪntəmət
+intimate_vrb	ˈɪntɪˌmeɪt
+intrigue_nou	ˈɪntɹiɡ
+intrigue_nou-vrb	ˌɪnˈtɹiɡ
+invalid_adj	ˌɪnˈvæɫəd
+invalid_nou	ˈɪnvəɫəd
+invert_adj-nou	ˈɪnvɝt
+invert_vrb	ənˈvɝt
+invite_nou	ˈɪnˌvaɪt
+invite_vrb	ənˈvaɪt
+isolate_nou	ˈaɪsəɫət
+isolate	ˈaɪsəˌɫeɪt
+jesus	ˈdʒizəs
+jesus_es	ˌheɪˈsus
+job_bible	ˈdʒoʊb
+job	ˈdʒɑb
+laminate_nou	ˈɫæmənət
+laminate_vrb	ˈɫæməˌneɪt
+lead_nou	ˈɫɛd
+lead_nou-vrb	ˈɫid
+learned_adj	ˈɫɝnɪd
+learned_vrb	ˈɫɝnd
+live_adj	ˈɫaɪv
+live_vrb	ˈɫɪv
+lives_nou	ˈɫaɪvz
+lives_vrb	ˈɫɪvz
+mate_nou	ˈmɑˌteɪ
+mate	ˈmeɪt
+minute_adj	maɪˈnut
+minute	ˈmɪnət
+misuse_nou	mɪsˈjus
+misuse_vrb	mɪsˈjuz
+mobile	ˈmoʊbəɫ
+mobile_geo	ˌmoʊˈbiɫ
+mobile_nou-art	ˈmoʊˌbiɫ
+moderate_adj-nou	ˈmɑdɝət
+moderate_vrb	ˈmɑdɝˌeɪt
+moped_nou	ˈmoʊpɛd
+moped_vrb	ˈmoʊpt
+mouth_nou	ˈmaʊθ
+mouth_vrb	ˈmaʊð
+nestle_nam	ˈnɛsˈɫi
+nestle_vrb	ˈnɛsəɫ
+object_nou	ˈɑbdʒɛkt
+object_vrb	əbˈdʒɛkt
+ornament_nou	ˈɔɹnəmənt
+ornament_vrb	ˌɔɹnəˈmɛnt
+overthrow_nou	ˈoʊvɝˌθɹoʊ
+overthrow_vrb	ˌoʊvɝˈθɹoʊ
+pasty_adj	ˈpeɪˌsti
+pasty_nou	ˈpæsˌti
+perfect_adj	ˈpɝˌfɪkt
+perfect_vrb	pɝˈfɛkt
+perfume_nou	ˈpɝˌfjum
+perfume_vrb	pɝˈfjum
+permit_nou	ˈpɝˌmɪt
+permit_vrb	pɝˈmɪt
+pervert_nou	ˈpɝvɝt
+pervert_vrb	pɝˈvɝt
+pigment_nou	ˈpɪɡmənt
+pigment_vrb	pəɡˈmɛnt
+polish_geo	ˈpoʊɫɪʃ
+polish	ˈpɑɫɪʃ
+postulate_nou	ˈpɑstʃəɫət
+postulate_vrb	ˈpɑstʃəˌɫeɪt
+precipitate_adj-nou	pɹəˈsɪpətət
+precipitate_vrb	pɹəˈsɪpəˌteɪt
+predicate_nou	ˈpɹɛdɪkət
+predicate_vrb	ˈpɹɛdəˌkeɪt
+present_adj-nou	ˈpɹɛzənt
+present_vrb	pɹiˈzɛnt
+produce_nou	ˈpɹoʊdus
+produce_vrb	pɹəˈdus
+progress_nou	ˈpɹɑˌɡɹɛs
+progress_vrb	pɹəˈɡɹɛs
+project_nou	ˈpɹɑdʒɛkt
+project_vrb	pɹɑˈdʒɛkt
+protest_nou	ˈpɹoʊˌtɛst
+protest_vrb	pɹəˈtɛst
+ravel_nam	ɹəˈvɛɫ
+ravel_nou	ˈɹævəɫ
+reading_en	ˈɹidɪŋ
+reading_geo	ˈɹɛdɪŋ
+read_past	ˈɹɛd
+read_present	ˈɹid
+rebel_nou	ˈɹɛbəɫ
+rebel_vrb	ɹɪˈbɛɫ
+record_nou	ˈɹɛkɝd
+record_vrb	ɹəˈkɔɹd
+recount_nou	ˈɹiˌkaʊnt
+recount_vrb	ˌɹiˈkaʊnt
+refund_nou	ˈɹiˌfənd
+refund_vrb	ɹɪˈfənd
+refuse_nou	ˈɹɛfˌjuz
+refuse_vrb	ɹəfˈjuz
+reject_nou	ˈɹidʒɛkt
+reject_vrb	ɹɪˈdʒɛkt
+rerelease	ˈɹiɹəˌɫis
+rerelease_vrb	ˌɹiɹəˈɫis
+resume_nou	ˈɹɛzəˌmeɪ
+resume_vrb	ɹiˈzum
+retard_nou	ˈɹitɑɹd
+retard_vrb	ɹɪˈtɑɹd
+rodeo_geo	ˌɹoʊˈdeɪˌoʊ
+rodeo	ˈɹoʊdiˌoʊ
+row_1	ˈɹoʊ
+row_2	ˈɹaʊ
+sake_jp	ˈsɑˌkeɪ
+sake	ˈseɪk
+separate_adj	ˈsɛpɝɪt
+separate_vrb	ˈsɛpɝˌeɪt
+sow_nou	ˈsaʊ
+sow	ˈsoʊ
+subject_adj-nou	ˈsəbdʒɪkt
+subject_vrb	səbˈdʒɛkt
+subordinate_adj-nou	səˈbɔɹdənət
+subordinate_vrb	səˈbɔɹdəˌneɪt
+supplement_nou	ˈsəpɫəmənt
+supplement_vrb	ˈsəpɫəmənt
+suspect_adj-nou	ˈsəsˌpɛkt
+suspect_vrb	səˈspɛkt
+syndicate_nou	ˈsɪndɪkət
+syndicate_vrb	ˈsɪndəˌkeɪt
+tear_nou	ˈtɪɹ
+tear_vrb	ˈtɛɹ
+transform_nou	ˈtɹænsfɔɹm
+transform	tɹænsˈfɔɹm
+transplant_nou	ˈtɹænsˌpɫænt
+transplant_vrb	tɹænsˈpɫænt
+transport_nou	ˈtɹænspɔɹt
+transport_vrb	tɹænˈspɔɹt
+upset_nou	ˈəpˌsɛt
+upset_vrb	əpˈsɛt
+uses_nou	ˈjusəz
+uses_vrb	ˈjuzəz
+use_nou	ˈjus
+use_vrb	ˈjuz
+wind_nou	ˈwɪnd
+wind_vrb	ˈwaɪnd
+winds_nou	ˈwɪndz
+winds_vrb	ˈwaɪndz
+wound_nou-vrb	ˈwund
+wound_vrb	ˈwaʊnd
